feat(controller): add infrastructure for legacy and hubble control plane
